backend/tests/UserApiTests/UserServiceTests.cs
using System.Threading.Tasks;
using AutoMapper;
using Moq;
using UserApi.Authorization;
using UserApi.Entities;
using UserApi.Helpers;
using UserApi.Models;
using UserApi.Repositories;
using Xunit;

namespace UserApiTests.Services
{
    public class UserServiceTests
    {
        private readonly Mock<IUserRepository> _mockUserRepository = new Mock<IUserRepository>();
        private readonly Mock<IJwtUtils> _mockJwtUtils = new Mock<IJwtUtils>();
        private readonly Mock<IMapper> _mockMapper = new Mock<IMapper>();
        private readonly Mock<IPasswordHasher> _mockPasswordHasher = new Mock<IPasswordHasher>();
        private readonly UserService _userService;

        public UserServiceTests()
        {
            _userService = new UserService(
                _mockUserRepository.Object,
                _mockJwtUtils.Object,
                _mockMapper.Object,
                _mockPasswordHasher.Object
            );
        }

        [Fact]
        public async Task RegisterUser_ShouldCreateUser_WithValidLengths()
        {
            // Arrange
            var registrationRequest = new RegisterUserRequest
            {
                FirstName = "Aseel",
                LastName = "Alqoud",
                Username = "aseel_alqoud",
                // other registration properties...
            };

            _mockUserRepository.Setup(x => x.GetUserByUsernameAsync(registrationRequest.Username)).ReturnsAsync((User?)null);

            _mockMapper.Setup(m => m.Map<User>(It.IsAny<RegisterUserRequest>()))
                       .Returns(new User()); // You need to set up the mapping based on your actual mapping logic.

            _mockPasswordHasher.Setup(p => p.HashPassword(It.IsAny<string>()))
                              .Returns((new byte[] { 1, 2, 3 }, new byte[] { 4, 5, 6 })); // Replace with your actual hashing logic.

            // Act
            var result = await _userService.RegisterUser(registrationRequest);

            // Assert
            Assert.NotNull(result);
            Assert.Equal(registrationRequest.FirstName, result.FirstName);
            Assert.Equal(registrationRequest.LastName, result.LastName);
            Assert.Equal(registrationRequest.Username, result.Username);
            // Add more assertions as needed based on the expected behavior
        }

        [Fact]
        public async Task RegisterUser_ShouldFail_WithInvalidPasswordLength()
        {
            // Arrange
            var registrationRequest = new RegisterUserRequest
            {
                FirstName = "Jake",
                LastName = "Doe",
                Username = "jake_doe",
                Password = "short", // Password with invalid length
                // other registration properties...
            };

            // Act and Assert
            await Assert.ThrowsAsync<InvalidPasswordException>(async () =>
            {
                await _userService.RegisterUser(registrationRequest);
            });

            // Add more assertions as needed based on the expected behavior
        }

        // Add more tests using Moq for other scenarios...

        // Existing code...
    }

    internal class RegisterUserRequest
    {
        public string FirstName { get; set; }
        public string LastName { get; set; }
        public string Username { get; set; }
        public string Password { get; set; }
    }

    // Example User service implementation
    public class UserService
    {
        private readonly IUserRepository _userRepository;
        private readonly IJwtUtils _jwtUtils;
        private readonly IMapper _mapper;
        private readonly IPasswordHasher _passwordHasher;

        public UserService(
            IUserRepository userRepository,
            IJwtUtils jwtUtils,
            IMapper mapper,
            IPasswordHasher passwordHasher)
        {
            _userRepository = userRepository;
            _jwtUtils = jwtUtils;
            _mapper = mapper;
            _passwordHasher = passwordHasher;
        }

        public async Task<User> RegisterUser(RegisterUserRequest registrationRequest)
        {
            // Your registration logic here...

            // For the sake of example, creating a new user using AutoMapper.
            var newUser = _mapper.Map<User>(registrationRequest);

            // Hash the password using the provided password hasher.
            var (hash, salt) = _passwordHasher.HashPassword(registrationRequest.Password);

            // Set the hashed password and salt to the user entity.
            newUser.PasswordHash = hash;
            newUser.PasswordSalt = salt;

            // Save the new user to the repository or database.
            await _userRepository.AddUser(newUser);

            // You may return the created user or other relevant information.
            return newUser;
        }

        // Other methods in your UserService class...
    }

    // Example PasswordHasher implementation
    public interface IPasswordHasher
    {
        (byte[] Hash, byte[] Salt) HashPassword(string password);
        (byte[] hash, byte[] salt) HashPassword(object password);
    }

    public class YourPasswordHasher : IPasswordHasher
    {
        private object salt;

        public (byte[] Hash, byte[] Salt) HashPassword(string password)
        {
            // Replace with your actual password hashing implementation.
            byte[] hash = // Compute the hash from the password
            byte[] salt = // Generate a random salt

#pragma warning disable CS8619 // Nullability of reference types in value doesn't match target type.
            return (hash, salt);
#pragma warning restore CS8619 // Nullability of reference types in value doesn't match target type.
        }

        public (byte[] hash, byte[] salt) HashPassword(object password)
        {
            throw new NotImplementedException();
        }
    }
}
